+++ backend-plugins/backend-base-plugin/src/main/kotlin/ru/itbasis/gradle/backend/BackendBasePlugin.kt
@@ -27,5 +27,4 @@ class BackendBasePlugin : AbstractPlugin() {
 			it.execute(target)
 		}
 	}
-
 }