+++ backend-plugins/backend-base-plugin/src/test/kotlin/ru/itbasis/gradle/backend/BackendBasePluginTest.kt
@@ -1,11 +1,12 @@
 package ru.itbasis.gradle.backend
 
 import io.kotest.core.spec.style.FunSpec
-import io.kotest.matchers.collections.shouldContain
+import io.kotest.matchers.collections.contain
 import io.kotest.matchers.collections.singleElement
 import io.kotest.matchers.should
 import io.kotest.matchers.shouldNotBe
 import ru.itbasis.gradle.common.tests.getAllDependencies
+import ru.itbasis.gradle.common.tests.getAllDependenciesAsRows
 import ru.itbasis.gradle.common.tests.initTestProject
 
 class BackendBasePluginTest : FunSpec(
@@ -14,19 +15,21 @@ class BackendBasePluginTest : FunSpec(
 			val project = initTestProject()
 			project.pluginManager.apply("ru.itbasis.gradle.backend-base-plugin")
 
-			project.plugins.getPlugin(BackendBasePlugin::class.java) shouldNotBe null
+			project.plugins.findPlugin(BackendBasePlugin::class.java) shouldNotBe null
 
 			val allDependencies = project.getAllDependencies()
+			project.getAllDependenciesAsRows()
 
 			allDependencies.map {
 				it.module
-			}.shouldContain("kotlin-stdlib-jdk8")
+			} should contain("kotlin-stdlib-jdk8")
+//			FIXME } should containAll("kotlin-stdlib-jdk8", "kotest-extensions-junit5")
 
 			allDependencies.filter {
 				it.group == "org.jetbrains.kotlin"
 			}.map {
 				it.version
-			}.toSet() should singleElement("1.4.30")
+			}.toSet() should singleElement("1.4.31")
 		}
 	}
 )