+++ root-module-plugin/src/main/kotlin/ru/itbasis/gradle/rootmodule/VersionToMarkdownPlugin.kt
@@ -1,7 +1,0 @@
-package ru.itbasis.gradle.rootmodule
-
-import org.gradle.api.Project
-import org.gradle.api.Plugin
-
-//class VersionToMarkdownPlugin : Plugin<Project> {
-//}